backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
accelerate=1.8.1=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.12.14=pypi_0
aiohttp-retry=2.9.1=pypi_0
aiosignal=1.4.0=pypi_0
aom=3.6.0=hcec6c5f_0
async-timeout=5.0.1=pypi_0
attrs=25.3.0=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h6c40b1e_6
c-ares=1.34.5=hf13058a_0
ca-certificates=2025.7.9=hbd8a1cb_0
cairo=1.16.0=h08824b9_6
certifi=2025.7.9=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.2.1=pypi_0
contourpy=1.3.2=pypi_0
cvzone=1.6.1=pypi_0
cycler=0.12.1=pypi_0
dav1d=1.2.1=h0dc2134_0
eigen=3.4.0=h1c7c39f_0
expat=2.7.1=h6d0c2b6_0
ffmpeg=6.1.1=hd2ef4e1_3
filelock=3.18.0=pypi_0
flask=3.1.1=pypi_0
flask-cors=6.0.1=pypi_0
fontconfig=2.14.1=h269ac6a_3
fonttools=4.58.5=pypi_0
freetype=2.10.4=h4cff582_1
fribidi=1.0.10=hbcb3906_0
frozenlist=1.7.0=pypi_0
fsspec=2025.5.1=pypi_0
gdk-pixbuf=2.42.10=h688160f_2
gettext=0.21.0=h4e8c18a_2
giflib=5.2.2=h10d778d_0
glib=2.84.2=h918fcf5_0
glib-tools=2.84.2=h918fcf5_0
graphite2=1.3.14=he9d5cce_1
gst-plugins-base=1.24.12=h25bb215_1
gstreamer=1.24.12=hccc03b2_1
gstreamer-orc=0.4.41=h9fdcad8_0
harfbuzz=10.2.0=h5b8c2d6_0
hdf5=1.14.5=h80586a8_2
hf-xet=1.1.5=pypi_0
huggingface-hub=0.33.2=pypi_0
icu=73.2=hf5e326d_0
idna=3.10=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.5.1=pypi_0
jpeg=9e=hb7f2c08_3
kiwisolver=1.4.8=pypi_0
krb5=1.20.1=h049b76e_0
lame=3.100=hb7f2c08_1003
leptonica=1.82.0=h7d8ae9a_3
lerc=4.0.0=hb486fe8_0
libabseil=20250127.0=cxx17_h548131f_0
libarchive=3.7.7=h4c66a04_0
libblas=3.9.0=17_osx64_openblas
libcblas=3.9.0=17_osx64_openblas
libcurl=8.12.1=h9bcc28a_0
libcxx=17.0.6=hf547dac_4
libdeflate=1.22=h00291cd_0
libedit=3.1.20230828=h6c40b1e_0
libev=4.33=h10d778d_2
libffi=3.4.4=hecd8cb5_1
libgfortran=5.0.0=10_4_0_h97931a8_28
libgfortran5=11.3.0=h9dfd629_28
libglib=2.84.2=heb90364_0
libiconv=1.18=h4b5e92a_1
liblapack=3.9.0=17_osx64_openblas
libnghttp2=1.57.0=h9beae6a_0
libogg=1.3.5=he3325bb_1
libopenblas=0.3.23=openmp_h429af6e_0
libopus=1.5.2=he3325bb_0
libpng=1.6.39=h6c40b1e_0
libpq=12.20=h184c1cd_0
libprotobuf=5.29.3=h66a43dd_1
librsvg=2.56.3=ha9ee93f_1
libssh2=1.11.1=h3a17b82_0
libtheora=1.1.1=hfdf4475_1006
libtiff=4.7.0=h2dfa3ea_0
libvorbis=1.3.7=h046ec9c_0
libvpx=1.13.1=hcec6c5f_0
libwebp=1.3.2=hc927772_1
libwebp-base=1.3.2=h46256e1_1
libxml2=2.13.8=h6070cd6_0
llvm-openmp=17.0.6=hdd4a2e0_0
lz4-c=1.9.4=hf0c8a7f_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.3=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.6.3=pypi_0
mysql=8.4.0=h5364baf_2
ncurses=6.4=hcec6c5f_0
networkx=3.4.2=pypi_0
numpy=1.26.4=pypi_0
opencv=4.10.0=py310h894815f_6
openh264=2.1.1=h8346a28_0
openjpeg=2.5.2=h2d09ccc_1
openssl=3.5.1=hc426f3f_0
packaging=25.0=pypi_0
pandas=2.3.1=pypi_0
pango=1.50.7=h986b0af_1
pcre2=10.42=h9b97e30_1
pillow=11.3.0=pypi_0
pip=25.1=pyhc872135_2
pixman=0.40.0=h9ed2024_1
playsound=1.3.0=pypi_0
propcache=0.3.2=pypi_0
psutil=7.0.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyjwt=2.10.1=pypi_0
pyparsing=3.2.3=pypi_0
python=3.10.18=hc958d9f_0
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.10=2_cp310
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
qtbase=6.7.2=h2d11ce5_0
readline=8.2=hca72f7f_0
regex=2024.11.6=pypi_0
requests=2.32.4=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.7.0=pypi_0
scipy=1.15.3=pypi_0
setuptools=78.1.1=py310hecd8cb5_0
six=1.17.0=pypi_0
sqlite=3.45.3=h6c40b1e_0
sympy=1.14.0=pypi_0
tesseract=5.2.0=h6d0c2b6_2
threadpoolctl=3.6.0=pypi_0
tk=8.6.14=h0a12a5f_1
tokenizers=0.21.2=pypi_0
torch=2.2.2=pypi_0
torchvision=0.17.2=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.53.1=pypi_0
twilio=9.6.5=pypi_0
typing-extensions=4.14.1=pypi_0
tzdata=2025.2=pypi_0
ultralytics=8.3.165=pypi_0
ultralytics-thop=2.0.14=pypi_0
urllib3=2.5.0=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py310hecd8cb5_0
xz=5.6.4=h46256e1_1
yarl=1.20.1=pypi_0
zlib=1.2.13=h4b97444_1
zstd=1.5.6=h138b38a_0
